feat(prd-003): Chrome MV3 capture extension with epoch-aligned telemetry
Implements PRD-003 (capture extension), all eight acceptance criteria, with
AC8 honestly qualified below. Repo gate green: biome clean across 169 files,
9 of 9 projects typecheck, 306 tests pass.

apps/extension (60 tests):
- MV3 manifest with tabCapture, offscreen, webNavigation, storage, scripting,
  and webRequest, each justified in docs/permissions-justification.md per
  ADR-010. Action click starts and stops capture on the active tab.
- Offscreen recorder: chunked MediaRecorder with timeslice, epoch anchor at
  onstart, chunked upload, and an AudioContext re-route so tab audio stays
  audible to the user during capture.
- Content-script telemetry for click, pointermove at 30 Hz, scroll, and
  masked input, converted to true epoch via performance.timeOrigin.
- Element sampler emitting Recorder-style fallback selectors (css, aria,
  text, xpath, pierce, with data-testid preferred) plus rect, role, name,
  innerText, viewport, DPR, and scroll offsets.
- Route detection from webNavigation plus MAIN-world history and Navigation
  API patches; state-change classification over a MutationObserver window
  with a DOM delta summary.
- Settle markers from a per-tab webRequest in-flight counter excluding
  websocket, SSE, and beacon traffic.
- Session finalizer emitting zod-validated events.jsonl and meta.json shaped
  directly from the @waggle/ir extension types, so ingest groups the stream
  into steps without reshaping any field. Toggleable click ripple overlay.

AC8 evidence and its boundary. Two things ran against real Chromium. The
built extension loads via --load-extension and its MV3 service worker
registers, proving the manifest and permissions are real and loadable. The
real unmodified content-script bundle then drives the canonical six-step
fixture walkthrough, with each click timestamp compared against an
independent wall clock: worst case 9.5 ms against the 50 ms budget,
reproduced independently by the orchestrator. What is not proven is an
actual tabCapture recording or byte-level correlation to decoded video
frames, because that needs a headed display. docs/ac8-e2e-runbook.md is the
manual path to close that gap, and the gap is recorded in the blocked
register in EXECUTION_LEDGER.md rather than being claimed as passing.

Input masking returns {length, masked} shapes only; no raw keystroke value
ever leaves the content script, per ADR-008.

+++ apps/extension/src/lib/events.ts
@@ -146,7 +146,12 @@ export const SessionVideoInfoSchema = z.strictObject({
   filename: z.string().min(1),
   mimeType: z.string().min(1),
   /** True epoch sampled at `MediaRecorder.onstart`: the video's frame-0 timestamp. */
-  anchorEpochMs: z.number().int().nonnegative(),
+  // Not `.int()`: this is `performance.timeOrigin + performance.now()`
+  // (offscreen/recorder.ts's MediaRecorder.onstart handler), which is
+  // sub-millisecond precision and legitimately fractional - unlike
+  // `startEpochMs` above, which is `Date.now()` in the background service
+  // worker and is always a whole millisecond.
+  anchorEpochMs: z.number().nonnegative(),
   durationMs: z.number().nonnegative(),
   chunkCount: z.number().int().nonnegative(),
 });
